+++ nodes/time-inject.js
@@ -84,54 +84,63 @@ module.exports = function (RED) {
         this.positionConfig = RED.nodes.getNode(config.positionConfig);
         // this.debug('initialize timeInjectNode ' + util.inspect(config, { colors: true, compact: 10, breakLength: Infinity }));
 
-        this.time = config.time;
-        this.timeType = config.timeType || 'none';
-        this.timeDays = config.timeDays;
-        this.timeOnlyOddDays = config.timeOnlyOddDays;
-        this.timeOnlyEvenDays = config.timeOnlyEvenDays;
-        this.timeMonths = config.timeMonths;
-        this.timeAltDays = config.timeAltDays;
-        this.timeAltOnlyOddDays = config.timeAltOnlyOddDays;
-        this.timeAltOnlyEvenDays = config.timeAltOnlyEvenDays;
-        this.timeAltMonths = config.timeAltMonths;
+        this.timeData = {
+            type: config.timeType || 'none',
+            value : config.time,
+            offsetType : config.offsetType,
+            offset : config.offset || config.timeOffset || 0,
+            multiplier : config.offsetMultiplier || config.timeOffsetMultiplier || 60,
+            next : true,
+            days : config.timeDays,
+            months : config.timeMonths,
+            onlyOddDays: config.timeOnlyOddDays,
+            onlyEvenDays: config.timeOnlyEvenDays
+        };
 
-        if (this.timeDays === '') {
+        if (!this.timeData.offsetType) {
+            this.timeData.offsetType = ((this.timeData.offset === 0) ? 'none' : 'num');
+        }
+        if (this.timeData.days === '') {
             throw new Error('No valid days given! Please check settings!');
         }
-        if (this.timeAltDays === '') {
-            throw new Error('No valid alternate days given! Please check settings!');
-        }
-        if (this.timeMonths === '') {
+        if (this.timeData.months === '') {
             throw new Error('No valid month given! Please check settings!');
         }
-        if (this.timeAltMonths === '') {
-            throw new Error('No valid alternate month given! Please check settings!');
-        }
-        if (this.timeOnlyEvenDays && this.timeOnlyOddDays) {
-            this.timeOnlyEvenDays = false;
-            this.timeOnlyOddDays = false;
-        }
-        if (this.timeAltOnlyEvenDays && this.timeAltOnlyOddDays) {
-            this.timeAltOnlyEvenDays = false;
-            this.timeAltOnlyOddDays = false;
+        if (this.timeData.onlyEvenDays && this.timeData.onlyOddDays) {
+            this.timeData.onlyEvenDays = false;
+            this.timeData.onlyOddDays = false;
         }
 
-        this.offset = config.offset || config.timeOffset || 0;
-        this.offsetType = config.offsetType;
-        if (!this.offsetType) { this.offsetType = ((this.offset === 0) ? 'none' : 'num'); }
-        this.offsetMultiplier = config.offsetMultiplier || config.timeOffsetMultiplier || 60;
+        this.timeAltData = {
+            type: config.timeAltType || 'none',
+            value : config.timeAlt || '',
+            offsetType : config.timeAltOffsetType,
+            offset : config.timeAltOffset || 0,
+            multiplier : config.timeAltOffsetMultiplier || 60,
+            next : true,
+            days : config.timeAltDays,
+            months : config.timeAltMonths,
+            onlyOddDays: config.timeAltOnlyOddDays,
+            onlyEvenDays: config.timeAltOnlyEvenDays
+        };
+        if (!this.timeAltData.offsetType) { this.timeAltData.offsetType = ((this.timeAltData.offset === 0) ? 'none' : 'num'); }
+
+        if (this.timeAltData.days === '') {
+            throw new Error('No valid alternate days given! Please check settings!');
+        }
+        if (this.timeAltData.months === '') {
+            throw new Error('No valid alternate month given! Please check settings!');
+        }
+        if (this.timeAltData.onlyEvenDays && this.timeAltData.onlyOddDays) {
+            this.timeAltData.onlyEvenDays = false;
+            this.timeAltData.onlyOddDays = false;
+        }
 
         this.property = config.property || '';
         this.propertyType = config.propertyType || 'none';
         this.propertyOperator = config.propertyCompare || 'true';
         this.propertyThresholdValue = config.propertyThreshold;
         this.propertyThresholdType = config.propertyThresholdType;
-        this.timeAlt = config.timeAlt || '';
-        this.timeAltType = config.timeAltType || 'none';
-        this.timeAltOffset = config.timeAltOffset || 0;
-        this.timeAltOffsetType = config.timeAltOffsetType;
-        if (!this.timeAltOffsetType) { this.timeAltOffsetType = ((this.timeAltOffset === 0) ? 'none' : 'num'); }
-        this.timeAltOffsetMultiplier = config.timeAltOffsetMultiplier || 60;
 
         this.recalcTime = (config.recalcTime || 2) * 3600000;
 
@@ -165,13 +174,20 @@ module.exports = function (RED) {
          * creates the timeout
          * @param {*} node - the node representation
          * @param {boolean} [_onInit] - _true_ if is in initialisation
+         * @param {Date} [dNow] - Date object with the calculation base
          * @returns {object} state or error
          */
-        function doCreateTimeout(node, _onInit) {
+        function doCreateTimeout(node, _onInit, dNow) {
+            node.debug(`doCreateTimeout _onInit=${_onInit}`);
             let errorStatus = '';
             let warnStatus = '';
             let isAltFirst = false;
             let isFixedTime = true;
+            let timeValid = (node.timeData && node.timeData.type !== 'none');
+            let timeAltValid = (node.timeAltData && node.propertyType !== 'none' && node.timeAltData.type !== 'none');
+            node.timeData.now = dNow || new Date();
+            node.timeAltData.now = node.timeData.now;
+
             node.nextTime = null;
             node.nextTimeAlt = null;
 
@@ -202,32 +218,21 @@ module.exports = function (RED) {
                     if (dNow < dStart || dNow > dEnd) {
                         node.nextTime = new Date(dStart.getFullYear() + ((dNow >= dEnd) ? 1 : 0), dStart.getMonth(), dStart.getDate(), 0, 0, 1);
                         warnStatus = RED._('time-inject.errors.invalid-daterange') + ' [' + node.positionConfig.toDateString(node.nextTime)+ ']';
-                        node.timeType = 'none';
-                        node.timeAltType = 'none';
+                        timeValid = false;
+                        timeAltValid = false;
                     }
                 } else {
                     // switch between year from end to start - e.g. 2.11. - 20.3.
                     if (dNow < dStart && dNow > dEnd) {
                         node.nextTime = new Date(dStart.getFullYear(), dStart.getMonth(), dStart.getDate(), 0, 0, 1);
                         warnStatus = RED._('time-inject.errors.invalid-daterange') + ' [' + node.positionConfig.toDateString(node.nextTime)+ ']';
-                        node.timeType = 'none';
-                        node.timeAltType = 'none';
+                        timeValid = false;
+                        timeAltValid = false;
                     }
                 }
             }
-            if (node.timeType !== 'none' && node.positionConfig) {
-                node.nextTimeData = node.positionConfig.getTimeProp(node, undefined, {
-                    type: node.timeType,
-                    value : node.time,
-                    offsetType : node.offsetType,
-                    offset : node.offset,
-                    multiplier : node.offsetMultiplier,
-                    next : true,
-                    days : node.timeDays,
-                    months : node.timeMonths,
-                    onlyOddDays: node.timeOnlyOddDays,
-                    onlyEvenDays: node.timeOnlyEvenDays
-                });
+            if (timeValid && node.positionConfig) {
+                node.nextTimeData = node.positionConfig.getTimeProp(node, undefined, node.timeData);
                 if (node.nextTimeData.error) {
                     errorStatus = 'could not evaluate time';
                     node.nextTime = null;
@@ -243,21 +248,8 @@ module.exports = function (RED) {
                 }
             }
 
-            if (node.propertyType !== 'none' &&
-                node.timeAltType !== 'none' &&
-                node.positionConfig) {
-                node.nextTimeAltData = node.positionConfig.getTimeProp(node, undefined, {
-                    type: node.timeAltType,
-                    value : node.timeAlt,
-                    offsetType : node.timeAltOffsetType,
-                    offset : node.timeAltOffset,
-                    multiplier : node.timeAltOffsetMultiplier,
-                    next : true,
-                    days : node.timeAltDays,
-                    months : node.timeAltMonths,
-                    onlyOddDays: node.timeAltOnlyOddDays,
-                    onlyEvenDays: node.timeAltOnlyEvenDays
-                });
+            if (timeAltValid && node.positionConfig) {
+                node.nextTimeAltData = node.positionConfig.getTimeProp(node, undefined, node.timeAltData);
 
                 if (node.nextTimeAltData.error) {
                     errorStatus = 'could not evaluate alternate time';
@@ -414,11 +406,12 @@ module.exports = function (RED) {
             // If this is pre-1.0, 'done' will be undefined
             done = done || function (text, msg) { if (text) { return node.error(text, msg); } return null; };
             send = send || function (...args) { node.send.apply(node, args); };
+            let dNow = new Date();
 
             try {
                 msg._srcid = node.id;
-                node.debug('--------- time-inject - input');
-                doCreateTimeout(node);
+                node.debug('--------- time-inject - input (type=' + msg.type + ')');
+                doCreateTimeout(node, false, dNow);
                 msg.topic = config.topic;
                 if (!node.positionConfig) {
                     throw new Error('configuration missing!');
@@ -430,7 +423,8 @@ module.exports = function (RED) {
                     offsetType: config.payloadOffsetType,
                     offset: config.payloadOffset,
                     multiplier: config.payloadOffsetMultiplier,
-                    next: true
+                    next: true,
+                    now: dNow
                 });
                 if (value === null || (typeof value === 'undefined')) {
                     throw new Error('could not evaluate ' + config.payloadType + '.' + config.payload);
@@ -452,7 +446,8 @@ module.exports = function (RED) {
                         offset: config.addPayload1Offset,
                         multiplier: config.addPayload1OffsetMultiplier,
                         next: config.addPayload1Next,
-                        days: config.addPayload1Days
+                        days: config.addPayload1Days,
+                        now: dNow
                     });
                     tsSetAddProp(this, msg, {
                         outType: config.addPayload2Type,
@@ -464,7 +459,8 @@ module.exports = function (RED) {
                         offset: config.addPayload2Offset,
                         multiplier: config.addPayload2OffsetMultiplier,
                         next: config.addPayload2Next,
-                        days: config.addPayload2Days
+                        days: config.addPayload2Days,
+                        now: dNow
                     });
                     tsSetAddProp(this, msg, {
                         outType: config.addPayload3Type,
@@ -476,7 +472,8 @@ module.exports = function (RED) {
                         offset: config.addPayload3Offset,
                         multiplier: config.addPayload3OffsetMultiplier,
                         next: config.addPayload3Next,
-                        days: config.addPayload3Days
+                        days: config.addPayload3Days,
+                        now: dNow
                     });
                 }
                 send(msg); // node.send(msg);
@@ -509,7 +506,7 @@ module.exports = function (RED) {
 
                 config.onceTimeout = setTimeout(() => {
                     node.emit('input', {
-                        type: 'once'
+                        type: 'once/startup'
                     });
                     doCreateTimeout(node);
                 }, (config.onceDelay || 0.1) * 1000);